+++ src/main/java/com/github/cstroe/spendhawk/web/user/UserController.java
@@ -10,8 +10,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import java.util.Optional;
-
 @Controller
 @RequestMapping("/user")
 public class UserController {
@@ -29,7 +27,7 @@ public class UserController {
     }
 
     @GetMapping("/{userId}")
-    public String view(@PathVariable Integer userId, Model model) {
+    public String view(@PathVariable Long userId, Model model) {
         User user = userRepository.findById(userId);
         if(user != null) {
             model.addAttribute("user", user);
